+++ packages/ethereum-contracts/contracts/agreements/ConstantFlowAgreementV1.sol
@@ -734,7 +734,7 @@ contract ConstantFlowAgreementV1 is
                 ? flowRateAllowance
                 : flowRateAllowance - flowRate;
             if (updatedFlowRateAllowance < 0) revert CFA_ACL_FLOW_RATE_ALLOWANCE_EXCEEDED();
-            _updateFlowRateAllowance(token, flowOperatorId, permissions, updatedFlowRateAllowance);
+            _updateFlowOperatorData(token, flowOperatorId, permissions, updatedFlowRateAllowance);
         }
         {
             _StackVars_createOrUpdateFlow memory flowVars;
@@ -783,7 +783,7 @@ contract ConstantFlowAgreementV1 is
                 ? flowRateAllowance
                 : flowRateAllowance - (flowRate - oldFlowData.flowRate);
             if (updatedFlowRateAllowance < 0) revert CFA_ACL_FLOW_RATE_ALLOWANCE_EXCEEDED();
-            _updateFlowRateAllowance(token, flowOperatorId, permissions, updatedFlowRateAllowance);
+            _updateFlowOperatorData(token, flowOperatorId, permissions, updatedFlowRateAllowance);
         }
 
         {
@@ -833,37 +833,7 @@ contract ConstantFlowAgreementV1 is
         int96 addedFlowRateAllowance, // flowRateBudget
         bytes calldata ctx
     ) public override returns (bytes memory newCtx) {
-        newCtx = ctx;
-
-        // [SECURITY] NOTE: we are holding the assumption here that ctx is correct and we validate it with
-        // authorizeTokenAccess:
-        ISuperfluid.Context memory currentContext = AgreementLibrary.authorizeTokenAccess(token, ctx);
-
-        if (currentContext.msgSender == flowOperator) revert CFA_ACL_NO_SENDER_FLOW_OPERATOR();
-        if (addedFlowRateAllowance < 0) revert CFA_ACL_NO_NEGATIVE_ALLOWANCE();
-        
-        (
-            bytes32 flowOperatorId,
-            uint8 oldPermissions,
-            int96 oldFlowRateAllowance
-        ) = getFlowOperatorData(token, currentContext.msgSender, flowOperator);
-
-        // @note this will revert if it overflows
-        int96 newFlowRateAllowance = oldFlowRateAllowance + addedFlowRateAllowance;
-        _updateFlowRateAllowance(
-            token,
-            flowOperatorId,
-            oldPermissions,
-            newFlowRateAllowance
-        );
-
-        emit FlowOperatorUpdated(
-            token,
-            currentContext.msgSender,
-            flowOperator,
-            oldPermissions,
-            newFlowRateAllowance
-        );
+        newCtx = increaseFlowRateAllowanceWithPermissions(token, flowOperator, 0, addedFlowRateAllowance, ctx);
     }
 
     /// @dev IConstantFlowAgreementV1.decreaseFlowRateAllowance implementation
@@ -873,41 +843,7 @@ contract ConstantFlowAgreementV1 is
         int96 subtractedFlowRateAllowance, // flowRateBudget
         bytes calldata ctx
     ) public override returns (bytes memory newCtx) {
-        newCtx = ctx;
-
-        // [SECURITY] NOTE: we are holding the assumption here that ctx is correct and we validate it with
-        // authorizeTokenAccess:
-        ISuperfluid.Context memory currentContext = AgreementLibrary
-            .authorizeTokenAccess(token, ctx);
-
-        if (currentContext.msgSender == flowOperator) revert CFA_ACL_NO_SENDER_FLOW_OPERATOR();
-        if (subtractedFlowRateAllowance < 0) revert CFA_ACL_NO_NEGATIVE_ALLOWANCE();
-
-        (
-            bytes32 flowOperatorId,
-            uint8 oldPermissions,
-            int96 oldFlowRateAllowance
-        ) = getFlowOperatorData(token, currentContext.msgSender, flowOperator);
-        
-        int96 newFlowRateAllowance = oldFlowRateAllowance - subtractedFlowRateAllowance;
-
-        // @note this defends against negative allowance
-        if (newFlowRateAllowance < 0) revert CFA_ACL_NO_NEGATIVE_ALLOWANCE();
-
-        _updateFlowRateAllowance(
-            token,
-            flowOperatorId,
-            oldPermissions,
-            newFlowRateAllowance
-        );
-
-        emit FlowOperatorUpdated(
-            token,
-            currentContext.msgSender,
-            flowOperator,
-            oldPermissions,
-            newFlowRateAllowance
-        );
+        newCtx = decreaseFlowRateAllowanceWithPermissions(token, flowOperator, 0, subtractedFlowRateAllowance, ctx);
     }
 
     /// @dev IConstantFlowAgreementV1.updateFlowOperatorPermissions implementation
@@ -920,7 +856,7 @@ contract ConstantFlowAgreementV1 is
     ) public override returns (bytes memory newCtx) {
         newCtx = ctx;
         ISuperfluid.Context memory currentContext =
-            _validateUpdateFlowOperatorDataInput(token, flowOperator, permissions, flowRateAllowance, ctx);
+            _validateAndAuthorizeUpdateFlowOperatorDataInput(token, flowOperator, permissions, flowRateAllowance, ctx);
         FlowOperatorData memory flowOperatorData;
         flowOperatorData.permissions = permissions;
         flowOperatorData.flowRateAllowance = flowRateAllowance;
@@ -930,65 +866,65 @@ contract ConstantFlowAgreementV1 is
         emit FlowOperatorUpdated(token, currentContext.msgSender, flowOperator, permissions, flowRateAllowance);
     }
 
-    /// @dev IConstantFlowAgreementV1.increaseFlowAllowanceWithPermissions implementation
-    function increaseFlowAllowanceWithPermissions(
+    /// @dev IConstantFlowAgreementV1.increaseFlowRateAllowanceWithPermissions implementation
+    function increaseFlowRateAllowanceWithPermissions(
         ISuperfluidToken token,
         address flowOperator,
         uint8 permissionsToAdd,
         int96 addedFlowRateAllowance,
         bytes calldata ctx
-    ) external override returns (bytes memory newCtx) {
+    ) public override returns (bytes memory newCtx) {
         newCtx = ctx;
-        ISuperfluid.Context memory currentContext =
-            _validateUpdateFlowOperatorDataInput(token, flowOperator, permissionsToAdd, addedFlowRateAllowance, ctx);
+        ISuperfluid.Context memory currentContext = _validateAndAuthorizeUpdateFlowOperatorDataInput(
+            token, flowOperator, permissionsToAdd, addedFlowRateAllowance, ctx
+        );
 
         (bytes32 flowOperatorId, uint8 oldPermissions, int96 oldFlowRateAllowance) =
             getFlowOperatorData(token, currentContext.msgSender, flowOperator);
 
-        FlowOperatorData memory flowOperatorData;
-        flowOperatorData.permissions = deltaAddPermissions(oldPermissions, permissionsToAdd);
         // @note this will revert if it overflows
         int96 newFlowRateAllowance = oldFlowRateAllowance + addedFlowRateAllowance;
-        flowOperatorData.flowRateAllowance = newFlowRateAllowance;
-        token.updateAgreementData(flowOperatorId, _encodeFlowOperatorData(flowOperatorData));
+        uint8 permissions = addPermissions(oldPermissions, permissionsToAdd);
+        _updateFlowOperatorData(
+            token, flowOperatorId, permissions, newFlowRateAllowance
+        );
 
         emit FlowOperatorUpdated(
-            token, currentContext.msgSender, flowOperator, flowOperatorData.permissions, newFlowRateAllowance
+            token, currentContext.msgSender, flowOperator, permissions, newFlowRateAllowance
         );
     }
 
-    /// @dev IConstantFlowAgreementV1.decreaseFlowAllowanceWithPermissions implementation
-    function decreaseFlowAllowanceWithPermissions(
+    /// @dev IConstantFlowAgreementV1.decreaseFlowRateAllowanceWithPermissions implementation
+    function decreaseFlowRateAllowanceWithPermissions(
         ISuperfluidToken token,
         address flowOperator,
         uint8 permissionsToRemove,
         int96 subtractedFlowRateAllowance,
         bytes calldata ctx
-    ) external override returns (bytes memory newCtx) {
+    ) public override returns (bytes memory newCtx) {
         newCtx = ctx;
-        ISuperfluid.Context memory currentContext = _validateUpdateFlowOperatorDataInput(
+        ISuperfluid.Context memory currentContext = _validateAndAuthorizeUpdateFlowOperatorDataInput(
             token, flowOperator, permissionsToRemove, subtractedFlowRateAllowance, ctx
         );
 
         (bytes32 flowOperatorId, uint8 oldPermissions, int96 oldFlowRateAllowance) =
             getFlowOperatorData(token, currentContext.msgSender, flowOperator);
 
-        FlowOperatorData memory flowOperatorData;
-        flowOperatorData.permissions = deltaRemovePermissions(oldPermissions, permissionsToRemove);
+        uint8 permissions = removePermissions(oldPermissions, permissionsToRemove);
         int96 newFlowRateAllowance = oldFlowRateAllowance - subtractedFlowRateAllowance;
 
         // @note this defends against negative allowance
         if (newFlowRateAllowance < 0) revert CFA_ACL_NO_NEGATIVE_ALLOWANCE();
-        flowOperatorData.flowRateAllowance = newFlowRateAllowance;
-
-        token.updateAgreementData(flowOperatorId, _encodeFlowOperatorData(flowOperatorData));
+        _updateFlowOperatorData(
+            token, flowOperatorId, permissions, newFlowRateAllowance
+        );
 
         emit FlowOperatorUpdated(
-            token, currentContext.msgSender, flowOperator, flowOperatorData.permissions, newFlowRateAllowance
+            token, currentContext.msgSender, flowOperator, permissions, newFlowRateAllowance
         );
     }
 
-    function deltaAddPermissions(uint8 existingPermissions, uint8 permissionDelta)
+    function addPermissions(uint8 existingPermissions, uint8 permissionDelta)
         public
         pure
         returns (uint8)
@@ -996,7 +932,7 @@ contract ConstantFlowAgreementV1 is
         return existingPermissions | permissionDelta;
     }
 
-    function deltaRemovePermissions(uint8 existingPermissions, uint8 permissionDelta)
+    function removePermissions(uint8 existingPermissions, uint8 permissionDelta)
         public
         pure
         returns (uint8)
@@ -1005,11 +941,11 @@ contract ConstantFlowAgreementV1 is
     }
     
     /// @dev This function ensures:
-    /// - token access
+    /// - token access is authorized
     /// - passed permissions are "clean"
     /// - no sender flow operator
     /// - no negative allowance
-    function _validateUpdateFlowOperatorDataInput(
+    function _validateAndAuthorizeUpdateFlowOperatorDataInput(
         ISuperfluidToken token,
         address flowOperator,
         uint8 permissions,
@@ -1017,8 +953,6 @@ contract ConstantFlowAgreementV1 is
         bytes calldata ctx
     ) internal returns (ISuperfluid.Context memory currentContext) {
         if (!FlowOperatorDefinitions.isPermissionsClean(permissions)) revert CFA_ACL_UNCLEAN_PERMISSIONS();
-        // [SECURITY] NOTE: we are holding the assumption here that ctx is correct and we validate it with
-        // authorizeTokenAccess:
         currentContext = AgreementLibrary.authorizeTokenAccess(token, ctx);
         if (currentContext.msgSender == flowOperator) revert CFA_ACL_NO_SENDER_FLOW_OPERATOR();
         if (flowAllowance < 0) revert CFA_ACL_NO_NEGATIVE_ALLOWANCE();
@@ -1130,17 +1064,19 @@ contract ConstantFlowAgreementV1 is
         return _decodeFlowOperatorData(uint256(data[0]));
     }
 
-    function _updateFlowRateAllowance
+    // @dev This function updates the flow operator data
+    // for `flowOperatorId` and sets it in token storage
+    function _updateFlowOperatorData
     (
         ISuperfluidToken token,
         bytes32 flowOperatorId,
-        uint8 existingPermissions,
+        uint8 updatedPermissions,
         int96 updatedFlowRateAllowance
     )
         private
     {
         FlowOperatorData memory flowOperatorData;
-        flowOperatorData.permissions = existingPermissions;
+        flowOperatorData.permissions = updatedPermissions;
         flowOperatorData.flowRateAllowance = updatedFlowRateAllowance;
         token.updateAgreementData(flowOperatorId, _encodeFlowOperatorData(flowOperatorData));
     }