+++ packages/ethereum-contracts/CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Unreleased
 
 ### Added
-- `increaseFlowAllowanceWithPermissions` and `decreaseFlowAllowanceWithPermissions` added to `ConstantFlowAgreementV1.sol`
+- `increaseFlowRateAllowanceWithPermissions` and `decreaseFlowRateAllowanceWithPermissions` added to `ConstantFlowAgreementV1.sol`
 
 ## [v1.7.1] - 2023-06-27